+++ src/test/java/com/iexec/sms/tee/TeeControllerTests.java
@@ -136,6 +136,7 @@ class TeeControllerTests {
     @Test
     void shouldGetSconeFramework() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -155,6 +156,7 @@ class TeeControllerTests {
     @Test
     void shouldGetGramineFramework() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );
@@ -175,6 +177,7 @@ class TeeControllerTests {
     @Test
     void shouldGetSconeProperties() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -203,6 +206,7 @@ class TeeControllerTests {
     @Test
     void shouldGetGramineProperties() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );
@@ -228,6 +232,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetSconePropertiesSinceGramineSms() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -247,6 +252,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetGraminePropertiesSinceSconeSms() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );
@@ -267,6 +273,7 @@ class TeeControllerTests {
     @Test
     void shouldGetSconePropertiesVersion() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -296,6 +303,7 @@ class TeeControllerTests {
     @Test
     void shouldGetGraminePropertiesVersion() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );
@@ -323,6 +331,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetSconePropertiesVersionSinceGramineSms() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -342,6 +351,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetSconePropertiesVersionSinceWrongVersion() {
         final TeeServicesProperties sconeProperties = new SconeServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties,
                 LAS_IMAGE
@@ -361,6 +371,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetGraminePropertiesVersionSinceSconeSms() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );
@@ -379,6 +390,7 @@ class TeeControllerTests {
     @Test
     void shouldNotGetGraminePropertiesVersionSinceWrongVersion() {
         final TeeServicesProperties gramineProperties = new GramineServicesProperties(
+                VERSION,
                 preComputeProperties,
                 postComputeProperties
         );